complete saved post functionality

+++ frontend/src/components/Posts/Posts.js
@@ -39,11 +39,14 @@ export default class Posts extends Component {
   }
 
   handleSave = (e) => {
-    let { loggedUser } = this.props;
-    if (e.target.className === "save-container") {
-      axios.post(`/api/users/${loggedUser.userData.id}/save`, {postId : e.target.id})
-    } else {
-      axios.delete(`/api/users/${loggedUser.userData.id}/save`, { data: {postId : e.target.id} }) //delete requests use config.data to add req.body
+    let { loggedUser, fetchUserSavedPosts } = this.props;
+    if (loggedUser) {
+      if (e.target.className === "save-container") {
+        axios.post(`/api/users/${loggedUser.userData.id}/save`, {postId : e.target.id})
+      } else {
+        axios.delete(`/api/users/${loggedUser.userData.id}/save`, { data: {postId : e.target.id} }) //delete requests use config.data to add req.body
+      }
+      fetchUserSavedPosts(loggedUser.userData.id)
     }
   }
